rename to ValidatorStore
