+++ node/validator_store.go
@@ -8,7 +8,6 @@ import (
 	"math/rand"
 	"os"
 	"path"
-	"strconv"
 	"sync"
 	"time"
 
@@ -19,6 +18,7 @@ import (
 	corev2 "github.com/Layr-Labs/eigenda/core/v2"
 	"github.com/Layr-Labs/eigenda/litt"
 	"github.com/Layr-Labs/eigenda/litt/littbuilder"
+	"github.com/Layr-Labs/eigenda/litt/types"
 	"github.com/Layr-Labs/eigenda/litt/util"
 	"github.com/Layr-Labs/eigensdk-go/logging"
 	"github.com/prometheus/client_golang/prometheus"
@@ -30,26 +30,21 @@ const (
 	BundleTableName          = "bundles"
 	MigrationTableName       = "migration"
 
-	// LevelDBPath is the path where the levelDB database is stored.
+	// LevelDBPath is he path where the levelDB database is stored.
 	LevelDBPath = "chunk_v2"
 	// LevelDBDeletionPath is the path where the levelDB database is stored while it is being deleted (for atomicity).
 	LevelDBDeletionPath = "chunk_v2_deleted"
 	// LittDBPath is the path where the littDB database is stored.
 	LittDBPath = "chunk_v2_litt"
-
-	// The name of the littDB table containing chunk data.
-	chunksTableName = "chunks"
-	// A legacy littDB table, this will exist until all old implementations are migrated and delete this table.
-	headersTableName = "headers"
-	// The metrics prefix for littDB.
-	littDBMetricsPrefix = "node_littdb"
 )
+
+// TODO consider renaming to ValidatorStore
 
 // BundleToStore is a struct that holds the bundle key and the bundle bytes.
 type BundleToStore struct {
 	// A bundle key, as encoded by BundleKey()
 	BundleKey []byte
-	// The binary bundle bytes.
+	// The bundle bytes, as encoded by TODO.
 	BundleBytes []byte
 }
 
@@ -68,9 +63,6 @@ type ValidatorStore interface {
 	// GetBundleData returns the chunks of a blob with the given bundle key.
 	// The returned chunks are encoded in bundle format.
 	GetBundleData(bundleKey []byte) ([]byte, error)
-
-	// Stop stops the store.
-	Stop() error
 }
 
 type validatorStore struct {
@@ -110,15 +102,11 @@ type validatorStore struct {
 
 	// The salt used to prevent an attacker from causing hash collisions in the duplicate request lock.
 	duplicateRequestSalt uint32
-
-	// A flag indicating whether the migration is complete. Used to prevent a double migration race condition
-	// (which is possible only in a unit test).
-	migrationComplete bool
 }
 
 var _ ValidatorStore = &validatorStore{}
 
-func NewValidatorStore(
+func NewStoreV2(
 	ctx context.Context,
 	logger logging.Logger,
 	config *Config,
@@ -136,11 +124,11 @@ func NewValidatorStore(
 	levelDBDeletionPath := path.Join(config.DbPath, LevelDBDeletionPath)
 
 	// If we previously made an attempt at deleting the levelDB database but it was interrupted, delete it now.
-	exists, err := util.Exists(levelDBDeletionPath)
-	if err != nil {
+	_, err := os.Stat(levelDBDeletionPath)
+	if err != nil && !os.IsNotExist(err) {
 		return nil, fmt.Errorf("failed to stat path %s: %v", levelDBDeletionPath, err)
 	}
-	if exists {
+	if err == nil {
 		// The previous attempt at deleting the levelDB database was interrupted.
 		logger.Warnf("partial deletion of levelDB database detected at %s. Deleting.", levelDBDeletionPath)
 		err = os.RemoveAll(levelDBDeletionPath)
@@ -194,9 +182,7 @@ func NewValidatorStore(
 		}
 	}
 
-	// Set up migration if necessary. The migrationComplete time is the time when all data in the old levelDB instance
-	// has exceeded its TTL. If the TTL is 2 weeks, then the migrationComplete time will be 2 weeks from now. At that
-	// moment, it becomes safe to permanently stop and delete the levelDB database.
+	// Set up migration if necessary.
 	var migrationComplete time.Time
 	if config.LittDBEnabled && levelDBExists {
 		// Both DBs are in play, meaning we are either about to start a migration or already in the middle of one.
@@ -207,16 +193,15 @@ func NewValidatorStore(
 		}
 		migrationKey := migrationKeyBuilder.Key([]byte("migrationCompleteTime"))
 
-		if !littDBExists {
+		if littDBExists {
 			// This is the first time we are starting up with littDB enabled and there is potentially data still
 			// in the levelDB. Start the data migration from levelDB to littDB.
 
 			migrationComplete = timeSource().Add(ttl)
-			logger.Infof("Beginning data migration from levelDB to littDB. Migration will be completed at %s",
+			logger.Infof("Begining data migration from levelDB to littDB. Migration will be completed at %s",
 				migrationComplete)
 
-			migrationCompleteUnix := migrationComplete.Unix()
-			err = levelDB.Put(migrationKey, []byte(fmt.Sprintf("%d", migrationCompleteUnix)))
+			err = levelDB.Put(migrationKey, []byte(fmt.Sprintf("%d", migrationComplete.Unix())))
 			if err != nil {
 				return nil, fmt.Errorf("failed to put migration key: %v", err)
 			}
@@ -228,11 +213,11 @@ func NewValidatorStore(
 			if err != nil {
 				return nil, fmt.Errorf("failed to get migration complete time: %v", err)
 			}
-			migrationCompleteUnix, err := strconv.ParseUint(string(migrationCompleteString), 10, 64)
+			migrationCompleteUnix, err := binary.ReadVarint(bytes.NewReader(migrationCompleteString))
 			if err != nil {
 				return nil, fmt.Errorf("failed to read migration complete time: %v", err)
 			}
-			migrationComplete = time.Unix(int64(migrationCompleteUnix), 0)
+			migrationComplete = time.Unix(migrationCompleteUnix, 0)
 
 			logger.Infof(
 				"Data migration from levelDB to littDB is in progress. Migration will be completed at %s",
@@ -248,37 +233,39 @@ func NewValidatorStore(
 	if config.LittDBEnabled {
 		logger.Infof("Using littDB at %s", littDBPath)
 
-		littConfig, err := litt.DefaultConfig(littDBPath)
+		littConfig, err := littbuilder.DefaultConfig(littDBPath)
 		littConfig.ShardingFactor = 1
 		littConfig.MetricsEnabled = true
 		littConfig.MetricsRegistry = registry
-		littConfig.MetricsNamespace = littDBMetricsPrefix
-		littConfig.Logger = logger
-		littConfig.DoubleWriteProtection = config.LittDBDoubleWriteProtection
+		littConfig.MetricsNamespace = "node_littdb"
 		if err != nil {
 			return nil, fmt.Errorf("failed to create new litt config: %w", err)
 		}
 
-		littDB, err = littbuilder.NewDB(littConfig)
+		littDB, err = littConfig.Build()
 		if err != nil {
 			return nil, fmt.Errorf("failed to create new litt store: %w", err)
 		}
 
-		chunkTable, err = littDB.GetTable(chunksTableName)
+		chunkTable, err = littDB.GetTable("chunks")
 		if err != nil {
 			return nil, fmt.Errorf("failed to get chunks table: %w", err)
 		}
 
-		// A prior implementation stored data here. Delete it if it exists.
-		// This is safe to delete once all old validators have been migrated to the new version.
-		err = littDB.DropTable(headersTableName)
+		headerTable, err = littDB.GetTable("headers")
 		if err != nil {
-			return nil, fmt.Errorf("failed to drop headers table: %w", err)
+			return nil, fmt.Errorf("failed to get headers table: %w", err)
 		}
 
 		err = chunkTable.SetTTL(ttl)
 		if err != nil {
 			return nil, fmt.Errorf("failed to set TTL for chunks table: %w", err)
+		}
+		// We store headers to prevent duplicate insertions. Store the headers for 2x the TTL to ensure that
+		// we don't delete the headers before the corresponding chunks.
+		err = headerTable.SetTTL(2 * ttl)
+		if err != nil {
+			return nil, fmt.Errorf("failed to set TTL for headers table: %w", err)
 		}
 	}
 
@@ -293,7 +280,7 @@ func NewValidatorStore(
 		headerTable:           headerTable,
 		ttl:                   ttl,
 		migrationCompleteTime: migrationComplete,
-		duplicateRequestLock:  common.NewIndexLock(1024),
+		duplicateRequestLock:  common.NewIndexLock(128),
 		duplicateRequestSalt:  rand.Uint32(),
 	}
 
@@ -311,7 +298,7 @@ func (s *validatorStore) StoreBatch(batchHeaderHash []byte, batchData []*BundleT
 	}
 
 	if s.littDB != nil {
-		size, err := s.storeBatchLittDB(batchData)
+		size, err := s.storeBatchLittDB(batchHeaderHash, batchData)
 		return nil, size, err
 	} else {
 		return s.storeBatchLevelDB(batchHeaderHash, batchData)
@@ -364,51 +351,75 @@ func (s *validatorStore) storeBatchLevelDB(batchHeaderHash []byte, batchData []*
 	return keys, size, nil
 }
 
-func (s *validatorStore) storeBatchLittDB(batchData []*BundleToStore) (uint64, error) {
+// storeBatchHeader stores the batch header hash in the database, returning an error if it is already present.
+// This method is guaranteed to only return nil exactly once if called multiple times with the same batch header hash.
+func (s *validatorStore) storeBatchHeaderHash(batchHeaderHash []byte) error {
+
+	// Grab a lock that is mutually exclusive for this batch header hash. This prevents us from storing
+	// data twice if we receive concurrent requests to store the same batch.
+	lockIndex := uint64(util.HashKey(batchHeaderHash[:], s.duplicateRequestSalt))
+	s.duplicateRequestLock.Lock(lockIndex)
+	defer s.duplicateRequestLock.Unlock(lockIndex)
+
+	ok, err := s.headerTable.Exists(batchHeaderHash[:])
+	if err != nil {
+		return fmt.Errorf("failed to check batch header existence: %v", err)
+	}
+	if ok {
+		return ErrBatchAlreadyExist
+	}
+
+	// Store batch header.
+	err = s.headerTable.Put(batchHeaderHash[:], []byte{})
+	if err != nil {
+		return fmt.Errorf("failed to put batch header: %v", err)
+	}
+
+	return nil
+}
+
+func (s *validatorStore) storeBatchLittDB(batchHeaderHash []byte, batchData []*BundleToStore) (uint64, error) {
 	var size uint64
 
-	writeCompleteChan := make(chan error, len(batchData))
+	// Store the batch header to prevent duplicate insertions.
+	err := s.storeBatchHeaderHash(batchHeaderHash)
+	if err != nil {
+		return 0, fmt.Errorf("failed to store batch header hash: %v", err)
+	}
+	size += uint64(len(batchHeaderHash))
+	err = s.headerTable.Flush()
+	if err != nil {
+		return 0, fmt.Errorf("failed to flush header table: %v", err)
+	}
+
+	// Now that the batch header is durable on disk, this validator will reject all future requests to store this batch.
+	// If the validator crashes between now and when it returns the availability signature, it will never
+	// return the availability signature for this batch. Although the validator could, in theory, reprocess this
+	// batch after starting back up, the performance and complexity overhead of doing so is cost prohibitive.
+	// Even if the validator was capable of reprocessing the batch after a crash, it's highly likely that the time
+	// window for returning the availability signature for this batch would have already passed -- thus negating
+	// any potential benefit of being able to reprocess the batch. It is, after all, not unreasonable to expect that
+	// a validator may not sign for some batches if it crashes.
+
+	// Store chunk data.
+	batch := make([]*types.KVPair, 0, len(batchData))
 	for _, batchDatum := range batchData {
 		bundleKeyBytes := batchDatum.BundleKey
 		bundleData := batchDatum.BundleBytes
-
-		go func() {
-			// Grab a lock on the hash of the blob. This protects against duplicate writes of the same blob.
-			lockIndex := uint64(util.HashKey(bundleKeyBytes[:], s.duplicateRequestSalt))
-			s.duplicateRequestLock.Lock(lockIndex)
-			defer s.duplicateRequestLock.Unlock(lockIndex)
-
-			exists, err := s.chunkTable.Exists(bundleKeyBytes[:])
-			if err != nil {
-				writeCompleteChan <- fmt.Errorf("failed to check existence: %v", err)
-				return
-			}
-
-			if exists {
-				// Data is already present, no need to write it again.
-				writeCompleteChan <- nil
-				return
-			}
-
-			err = s.chunkTable.Put(bundleKeyBytes, bundleData)
-			if err != nil {
-				writeCompleteChan <- fmt.Errorf("failed to put data: %v", err)
-			}
-
-			writeCompleteChan <- nil
-		}()
-
 		size += uint64(len(bundleKeyBytes) + len(bundleData))
+
+		batch = append(batch, &types.KVPair{
+			Key:   bundleKeyBytes,
+			Value: bundleData,
+		})
 	}
 
-	for i := 0; i < len(batchData); i++ {
-		err := <-writeCompleteChan
-		if err != nil {
-			return 0, err
-		}
+	err = s.chunkTable.PutBatch(batch)
+	if err != nil {
+		return 0, fmt.Errorf("failed to put batch: %v", err)
 	}
 
-	err := s.chunkTable.Flush()
+	err = s.chunkTable.Flush()
 	if err != nil {
 		return 0, fmt.Errorf("failed to flush chunk table: %v", err)
 	}
@@ -500,12 +511,6 @@ func (s *validatorStore) finalizeMigration(ctx context.Context) {
 		s.migrationLock.Lock()
 		defer s.migrationLock.Unlock()
 
-		if s.migrationComplete {
-			s.logger.Info("migration already completed, nothing to do")
-			return
-		}
-		s.migrationComplete = true
-
 		s.logger.Infof("migration to littDB complete, deleting levelDB at %s", s.levelDBPath)
 
 		err := s.levelDB.Shutdown()
@@ -539,21 +544,4 @@ func BundleKey(blobKey corev2.BlobKey, quorumID core.QuorumID) ([]byte, error) {
 		return nil, err
 	}
 	return buf.Bytes(), nil
-}
-
-func (s *validatorStore) Stop() error {
-	if s.littDB != nil {
-		err := s.littDB.Close()
-		if err != nil {
-			return fmt.Errorf("failed to close littDB: %v", err)
-		}
-	}
-	if s.levelDB != nil {
-		err := s.levelDB.Shutdown()
-		if err != nil {
-			return fmt.Errorf("failed to close levelDB: %v", err)
-		}
-	}
-
-	return nil
 }
